refactor: 기본 pk 타입 변경, 삭제시 수정일자 업데이트, 업무 목록 조회 jpa->mybatis 변경

+++ src/main/java/com/douzone/prosync/task_status/entity/TaskStatus.java
@@ -20,7 +20,7 @@ public class TaskStatus {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer taskStatusId;
+    private Long taskStatusId;
 
     private String taskStatus;
